+++ pytorch_segmentation_models_trainer/dataset_loader/__init__.py
@@ -1,3 +1,6 @@
 from pytorch_segmentation_models_trainer.dataset_loader.raster_patch_dataset import (  # noqa: F401
     RasterPatchDataset,
 )
+from pytorch_segmentation_models_trainer.dataset_loader.embedding_dataset import (  # noqa: F401
+    EmbeddingDataset,
+)